hint serialization fix
